visualisation refactored, new timeline option with occurrence threshold, visualiser classmethods finished

+++ cli/analyse_openings.py
@@ -6,6 +6,7 @@ from next_move.opening_directory import EcoDB
 
 # from next_move.visualiser.visualiser import Visualiser
 
+from pathlib import Path
 from tqdm import tqdm
 from typing import Optional
 
@@ -13,33 +14,36 @@ import click
 
 
 def run_analysis(player_id: str) -> Tree:
-    tree = Tree()
-    chess_com = ChessCom()
-    # visualiser = Visualiser()
-    games = chess_com.get_all_games(player_id)
-    stockfish = Stockfish("16/bin/stockfish")
-    eco_db = EcoDB("eco/openings.json")
+    if Path(f"next_move/cache/trees/{player_id}.json").exists():
+        return Tree.from_json(f"next_move/cache/trees/{player_id}.json")
+    else:
+        tree = Tree()
+        chess_com = ChessCom()
+        # visualiser = Visualiser()
+        games = chess_com.get_all_games(player_id)
+        stockfish = Stockfish("16/bin/stockfish")
+        eco_db = EcoDB("eco/openings.json")
 
-    game_processor = GameProcessor(tree, stockfish, eco_db, player_id)
+        game_processor = GameProcessor(tree, stockfish, eco_db, player_id)
 
-    for game in tqdm(games):
-        game_processor.process_game(game)
+        for game in tqdm(games):
+            game_processor.process_game(game)
 
-    stockfish.quit()  # it's fine for now, but will need to refactor this into a context manager
+        stockfish.quit()  # it's fine for now, but will need to refactor this into a context manager
 
-    # visualiser.sankey(
-    #     **game_processor.tree.to_sankey(prune_below_count=5), path="tree.html"
-    # )
-    # visualiser.timeline(
-    #     game_processor.tree.to_timeline(breakdown="M"), path="timeline.png"
-    # )
+        # visualiser.sankey(
+        #     **game_processor.tree.to_sankey(prune_below_count=5), path="tree.html"
+        # )
+        # visualiser.timeline(
+        #     game_processor.tree.to_timeline(breakdown="M"), path="timeline.png"
+        # )
 
-    # df = tree.to_opening_strength()
-    # df.to_csv("opening_strength.csv")
+        # df = tree.to_opening_strength()
+        # df.to_csv("opening_strength.csv")
 
-    tree.to_json(f"next_move/cache/trees/{player_id}.json")
+        tree.to_json(f"next_move/cache/trees/{player_id}.json")
 
-    return tree
+        return tree
 
 
 @click.command()
